+++ ontimize/components/list/renderers/o-list-item-card-renderer.class.ts
@@ -1,9 +1,4 @@
-import {
-  Injector,
-  ElementRef,
-  Renderer,
-  EventEmitter
-} from '@angular/core';
+import { ElementRef, EventEmitter, Injector, Renderer } from '@angular/core';
 
 import { OListItemComponent } from '../list-item/o-list-item.component';
 
@@ -11,13 +6,14 @@ export const DEFAULT_INPUTS_O_CARD_RENDERER = [
   'title',
   'subtitle',
   'image',
-  'action1Text : action-1-text',
-  'action2Text : action-2-text'
+  'showImage: show-image',
+  'action1Text: action-1-text',
+  'action2Text: action-2-text'
 ];
 
 export const DEFAULT_OUTPUTS_O_CARD_RENDERER = [
-  'onAction1Click : action-1',
-  'onAction2Click : action-2'
+  'onAction1Click: action-1',
+  'onAction2Click: action-2'
 ];
 
 export class OListItemCardRenderer {
@@ -29,6 +25,7 @@ export class OListItemCardRenderer {
   protected _title: string;
   protected _subtitle: string;
   protected _image: string;
+  protected _showImage: boolean = true;
   protected _action1Text: string;
   protected _action2Text: string;
 
@@ -41,8 +38,7 @@ export class OListItemCardRenderer {
     protected _renderer: Renderer,
     protected _injector: Injector,
     protected _listItem: OListItemComponent
-  ) {
-  }
+  ) { }
 
   modifyMatListItemElement() {
     if (this.elRef.nativeElement && this.elRef.nativeElement.parentElement) {
@@ -88,6 +84,14 @@ export class OListItemCardRenderer {
     this._image = val;
   }
 
+  get showImage(): boolean {
+    return this._showImage;
+  }
+
+  set showImage(val: boolean) {
+    this._showImage = val;
+  }
+
   get action1Text(): string {
     return this._action1Text;
   }
@@ -103,4 +107,5 @@ export class OListItemCardRenderer {
   set action2Text(val: string) {
     this._action2Text = val;
   }
+
 }