+++ ontimize/components/table/o-table.dao.ts
@@ -29,7 +29,7 @@ export class OTableDao {
   }
 
   removeQuery(filters: any): Observable<any> {
-    return Observable.merge(filters.map((kv => this.dataService[this.methods.delete](kv, this.entity))));
+    return Observable.merge(...filters.map((kv => this.dataService[this.methods.delete](kv, this.entity))));
   }
 
   insertQuery(av: Object, sqlTypes?: Object): Observable<any> {